adding 'depend_on_asset' for sprockets
http://rubydoc.info/gems/sprockets/2.10.1/file/README.md#The_depend_on_asset_Directive

If a dependency is used in an ERB asset that references another asset, it will not be updated when the reference asset is updated. The fix is to use //= depend_on or its cousin //= depend_on_asset however this is easy to forget. See rails/sprockets-rails#95 for more information.

Currently Rails/Sprockets hides this problem, and only surfaces it when the app is deployed with precompilation to production multiple times. We know that you will have this problem if you are referencing assets from within other assets and not declaring them as dependencies.
(text taken from https://github.com/schneems/sprockets_better_errors)